Graduate classes in supervision example
