+++ 2017-03_AkkaIntro/code/src/Supervision/AMonkeyGotBumped.cs
@@ -1,0 +1,14 @@
+﻿using Akka.Actor;
+
+namespace Supervision
+{
+    public class AMonkeyGotBumped
+    {
+        public IActorRef MonkeyThatGotBumped { get; }
+
+        public AMonkeyGotBumped(IActorRef bumpedMonkey)
+        {
+            MonkeyThatGotBumped = bumpedMonkey;
+        }
+    }
+}
