+++ 2017-03_AkkaIntro/code/src/Supervision/BumpedHeadException.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Runtime.Serialization;
+using Akka.Actor;
+
+namespace Supervision
+{
+    [Serializable]
+    public class BumpedHeadException : Exception
+    {
+        public IActorRef Monkey { get; }
+
+        public BumpedHeadException(IActorRef monkey)
+        {
+            Monkey = monkey;
+        }
+        protected BumpedHeadException(SerializationInfo info, StreamingContext context) : base(info, context)
+        {
+        }
+    }
+}
