+++ 2017-03_AkkaIntro/code/src/Supervision/CallTheDoctor.cs
@@ -1,0 +1,4 @@
+namespace Supervision
+{
+    public class CallTheDoctor { }
+}
