+++ 2017-03_AkkaIntro/code/src/Supervision/MonkeyCountUpdate.cs
@@ -1,0 +1,12 @@
+﻿namespace Supervision
+{
+    public class MonkeyCountUpdate
+    {
+        public int NumberMonkeys { get; }
+
+        public MonkeyCountUpdate(int numMonkeys)
+        {
+            NumberMonkeys = numMonkeys;
+        }
+    }
+}
